Builds a PubsubMessage Object

Builds a PubsubMessage Object

PubsubMessage(
  data = NULL,
  message_id = NULL,
  ordering_key = NULL,
  attributes = NULL,
  publish_time = NULL
)

Arguments

data

character The message data field as a base64 encoded string

message_id

`character ID of this message, assigned by the server when the message is published

ordering_key

character If non-empty, identifies related messages for which publish order should be respected

attributes

list Key-value pairs attributes for this message

publish_time

character The time at which the message was published, populated by the server when it receives the Publish call

Value

PubsubMessage object

See also