Modify the ack deadline for a subscription

This method is useful to indicate that more time is needed to process a message by the subscriber, or to make the message available for redelivery if the processing was interrupted.

Usage,
subscriptions_modify_ack_deadline(subscription, ack_ids, ack_deadline)

Arguments

subscription

character, Subscription A subscription name or Subscription object

ack_ids

character A vector containing ackIDs. They can be acquired using

ack_deadline

numeric The new ack deadline (in seconds)

Value

logical TRUE if successfully modified

See also