Updates an existing subscription.

Certain properties of a subscription, such as its topic, are not modifiable.

subscriptions_patch(
  subscription,
  topic,
  labels = NULL,
  dead_letter_policy = NULL,
  msg_retention_duration = NULL,
  retry_policy = NULL,
  push_config = NULL,
  ack_deadline = NULL,
  expiration_policy = NULL,
  filter = NULL,
  detached = NULL,
  retain_acked_msgs = NULL,
  enable_ordering = NULL
)

Arguments

subscription

character, Subscription Required, a subscription name or a Subscription object

topic

character, Topic Required, a topic name or a Topic object

labels

labels Key value pairs

dead_letter_policy

DeadLetterPolicy A DeadLetterPolicy object

msg_retention_duration

numeric How long to retain unacknowledged messages (in seconds)

retry_policy

RetryPolicy policy that specifies how Pub/Sub retries message delivery for this subscription, can be built with RetryPolicy

push_config

PushConfig Can be built with PushConfig

ack_deadline

numeric amount of time (in seconds) Pub/Sub waits for the subscriber to acknowledge receipt before resending the message

expiration_policy

ExpirationPolicy specifies the conditions for this subscription's expiration. Can be built with ExpirationPolicy

filter

character An expression written in the Pub/Sub filter language

detached

logical Indicates whether the subscription is detached from its topic

retain_acked_msgs

logical Indicates whether to retain acknowledged messages

enable_ordering

logicalmessages published with the same orderingKey in PubsubMessage will be delivered to the subscribers in the order in which they are received by the Pub/Sub system

Value

An updated Subscription object

See also